Extract album art palette extraction into util/AlbumArtPalette.kt
extractColorsFromArt was 60 lines of pure image processing inside
SpotifyViewModel. Extracted into a standalone suspend function that
takes a Context + URL and returns a ThemeColors. The ViewModel now
assigns the result directly. No behavior change.

Closes #165
